At level 100, players gain a 25x multiplier to damage done versus MoP bosses; as a result, I believe the following items are very likely to be soloable, so I've preemptively moved them to the solo section:
- Mounts: Clutch of Ji-Kun, Son of Galleon's Saddle, Spawn of Horridon, Reins of the Astral Cloud Serpent, Reins of the Cobalt Primordial Direhorn, Reins of the Heavenly Onyx Cloud Serpent, Reins of the Thundering Cobalt Cloud Serpent
- Pets: Dark Quivering Blob, Pygmy Direhorn, Quivering Blob, Droplet of Y'Shaarj, Gooey Sha-ling, Kovok

Kor'kron Juggernaut is a Mythic drop, and I'm not sure if it will be soloable at 100, so I've modified its group size to 2.

Added defeat detection for the following pets. Note that Rarity doesn't check that you defeated it on the right difficulty; that's up to you to ensure: Ji-Kun Hatchling, Dark Quivering Blob, Pygmy Direhorn, Quivering Blob, Blackfuse Bombling, Droplet of Y'Shaarj, Gooey Sha-ling, Kovok, Son of Animus

git-svn-id: https://repos.wowace.com/wow/rarity@374 6a7e8290-38a7-46e8-902e-0a51f977da95