Added filter system for the NPC tooltips
This will allow filtering of the extended tooltip info added to NPC
tooltips. It is based on predefined filter types, such as profession,
class, level, or defeated status (NYI).

Only the IS_SPELL_KNOWN type is currently supported, but the system will
likely be extended. Some more generalization is needed. It shall be
added in the future (when new filters are required).