Fixed attempts counter being cut off for very large numbers
Apparently some people have a LOT of attempts racked up, leading to them
being set to the lower value. I did not expect this to happen.

Therefore, I'm setting it to the maximum value that is supported by
string.format with the given %d (digit) placeholder. I don't want to
change all the localization phrases and use a different one, so this
remains the easier solution.