DB: Added lockout details for the G.M.O.D. mount
This allows resolving the defeat detection based on different bosses and
instance difficulties. In order to mark this mount as defeated, both
Jaina (in LFR) and Mekkatorque (all other difficulties) would have to be
defeated.

Previously, the item was marked as defeated as soon as Mekkatorque on
any difficulty was defeated, which is not only inaccurate (you could
still kill him on the other difficulties and get the mount), as well as
wrong (after the hotfix, he doesn't drop it in LFR, but there was no way
to differentiate those encounters or include another encounter for the
same item).