Core: Improved lockout-based defeat detection for shared difficulties
The previous version didn't work... I think. It was originally added for
the G.M.O.D. mount but I never got around to testing it (as that'd
require a tremendous effort on my part). With the MOP raids and LFR,
testing has become feasible, so it's a bit less experimental now...