Strip the color code from tooltip strings
This is not usually a problem, as most nodes are not color coded
(although I am not sure if any addons do modify this?). However,
with the addition of the Dirty Glinting Object, this produces an NPC
tooltip. For some users, this will be color coded and therefore it
will not properly detect the attempt on Lucy's Lost Collar.

The color coding will now be stripped if it does exist.