Core: Implemented a more streamlined Profiling API
Relying on debugprofilestart directly is a bit risky, and the existing Profiling code only supports two groups. In order to debug the performance issues more easily, a more powerful profiling API should prove helpful.

We can later remove the legacy profiling code, or gradually transition to the streamlined API. This one is based on the Chromium/Firefox console APIs and can be nested indefinitely as long as unique labels are used.

Cleanup: Removed isTimerRunning property from Profling API

It's unused and should never have been added in the first place.