UI: Fixed outdated check for relevant unit types
This hasn't been working since approximately 2014, when the GUID format appears to have changed. The effect of this broken early-exit condition is that zonewide items aren't displayed in the GameTooltip.

The one thing I'm not sure about here is the enum values:

It appears 3 is GameObject and 5 Vignette, but the logic doesn't make any sense with these values. So I'm assuming this is an oversight and the only relevant types - Pet and Creature - should be used.

The test is a formality and only exists because I wanted to be sure the pattern is actually working as expected.