UI: Added an early exit to avoid errors from outdated addons
It has been reported that the GetUnit call causes errors for some people.

I verified that Blizzard's GameTooltip uses the new GameTooltipDataMixin which includes this API, so the only explanation is that it's a tooltip created by another addon which isn't yet using this method.

It's quite a hacky "fix", but until I know what exactly is going on in these instances it should at least prevent the errors.