UI: Added another early exit for third-party tooltips
Whenever another addon makes use of a tooltip (frame that inherits GameTooltipDataMixin), multiple calls of the post-hook can add duplicate lines to the global GameTooltip.

I still can't test this since the issue doesn't manifest for me, but it should prevent this particular problem IFF my assumptions are correct. And if they aren't... well, then this is just useless but not actively harmful.