Core: Fixed detection for opening faction-specific containers
If a faction-specific item is obtained from a container, such as the mounts from Time Rift boxes, then the detection adds an attempt for it even though the player's faction doesn't match the requirement. This fix is "hacky" because I don't dare to touch the nested loops right now; ideally it'd be tested via CI and probably restructured a bit, but that will just have to wait (unfortunately).