DB: Removed Foot Ball
As the comment indicates, this can be purchased from a vendor (for gold). It doesn't seem very useful to track this item, and indeed may be aggravating players if the achievement pops up for a "worthless" collectible instead of one of the more desirable items.

Just noticed this comment while looking through the database, so I figured we might as well remove it.