UI: Fixed the tooltip for faction specific container items
As a follow-up to the attempts detection fix, the UI should reflect that the item can't be obtained by players of the opposite faction.

Example: Time Rift boxes display the Horde mount (wolf) on Alliance, but opening the box correctly doesn't add the attempts. That'll be quite confusing and it's best to just hide the item for Alliance players.