Cleanup: Removed the obsolete session locking mechanism
This was clearly an ill-conceived feature from the start (due to lack of experience on my part), only necessitated by the second LOOT_READY that Blizzard seemingly introduced after the loot window is opened.

Thanks to the bucketed sequence we can completely purge it and it still works, at least as long as events trigger within the same 0.5sec burst delay. Of course, this doesn't fix attempts being lost if players are looting too fast, but if that turns out to be a problem it's trivial to adjust the delay (or even make it configurable).