DB: Updated Everlasting Horn of Lavaswimming
It seems that the drop rate is questionably low and no information about this item was made available by Blizzard. Since I don't want to have people try and farm this repeatedly thinking they'll get it in 40 attempts, let's make it very clear that that the drop rate might be abysmal. Hopefully, this will prompt some research so people can make a more informed decision, or postpone farming this until more data has been gathered by the community.

In the same vein, using the weekly quest for defeat detection makes the most sense, to discourage players from repeatedly farming it (without doing more research first, that is). And to be even clearer, I'll also adjust the sourceText, although it may have to be updated again later should we get better info on this thing.