Repo: Restored the original luacheck configuration setup
Not sure what the idea was when I renamed it; I guess the error messages from luacheck analyzing its own config were annoying... But IDE support for nonstandard config files is an issue; you have to manually configure it for every workspace in vscode, for example.

Since this also allows running luacheck without args in the project root instead of using a dedicated script, which is easier to understand, I guess it's better to just go back to the original (standard) setup.