DB: Removed the statistic ID for Quantum Courser
The boss appears in two dungeons, and this statistic is for the alternate version (which doesn't drop the mount). Unfortunately, there's no statistic for the "right" version that could be used.

This fixes the tracking for Quantum Courser, which has been reported as broken by multiple people. It seems that I originally added the wrong statistic because wowhead didn't display the dungeon name? Or maybe I misread, who knows.