Core: Implemented support for Primal Storms via Area POIs
This is a more reliable way of tracking the Elemental Storm events, although it isn't quite perfect:

Firstly, the persistent elementals will be ignored if the storm that originally spawned them is no longer active. Since it's unknown whether they even drop the pets in this case and they won't respawn once killed anyway, I'd consider this a minor flaw that can probably be ignored (leading to a few missing attempts at worst).

Unfortunately, I haven't found a way of detecting whether the current area is affected. So Rarity simply checks whether any Elemental Storm is active in the current zone, which could lead to false positives if someone was to go to another area that can also be imbued (but currently isn't). However... who does that while farming the pets, anyway? For the most part this approach should work well, and if not then I'll have to think of a workaround for this problem later.