Core: Fixed various script errors on Classic Era realms
The tooltip hooking functionality is going to be disabled, which isn't ideal. It should be possible to adapt the original retail code and make both methods work alongside each other. Since I don't have time to look into that right now and it's probably not a critical issue (at least for the next release), I'll leave it for later. Classic Era likely needs more work in general.