Options: Explicitly set ChatFrame as the default output sink
This used to be the standard option in LibSink, but since Blizzard's UI revamp in 10.0, LibSink falls back to the floating combat text instead. There have been numerous reports of people not liking this change, or in some cases even being unable to see announcements at all. They will usually blame Rarity and/or be confused by the lack of selected output channels.