UI: Enable deferred loading of the Options UI
Currently, the module is either loaded (settings are registered) or not (settings not registered = menu entry missing).

With this change, the AceConfig registry setup is moved into the addon core. Ace will then be able to register the settings properly, even if the options aren't yet loaded. This not only fixes the glitch that had users confused, but also improves the "perceived" snappiness of opening the UI because most of the time is spent in Blizzard's menu code.

This cost is already paid if the UI has been opened before, whereas loading the options, registering the settings tree, then creating it, and THEN opening it leads to a noticeable delay (which is likely to be blamed on Rarity, even though it's just slow Lua-C-API calls in the client's native code).