UI: Implement a fallback to display when Options are disabled
With the lazy-loading approach, it's possible that players might disable the Options module (accidentally, usually) and then get script errors when AceConfig tries to load the UI tree.

This is a pretty basic fallback UI which allows loading the options seamlessly in case they were accidentally disabled. Don't know if there are other cases where this might fail, but it's better than nothing.