Refactor: Streamline the existing debug mode toggles
For debugging purposes, it would be useful to track various ingame events that aren't required for regular gameplay activities. If the addon always registers for them, some overhead would be introduced without any benefit for the average player. Having a fixed point for setup and teardown code that's only used for development or debugging makes it easier to handle situations like this gracefully.