Core: Fix luckiness calculation for repeated attempts
When an item has already been obtained, only the current run of attempts should be considered to determine "luckiness".
That is to say, once you find an item the percentage should be reset and not "remember" the total attempts count.

This behavior is consistent with the way that the LDB and main window counterparts work, as well as player expectations.