CI: Replace the luacheck setup with a curl command
The workflow is currently broken because LuaRocks' package registry seems to serve a manifest that's too large for LuaJIT to process. There are various workarounds, the most obvious one being to replace LuaJIT with PUC's Lua interpreter. But the more I read about this problem, the more I'm convinced that I don't want to rely on the LuaRocks ecosystem for this particular task.