Core: Remove tracking of "rare" achievement NPCs
The feature doesn't seem in line with what players use Rarity for these days, and those "rare" NPCs haven't been actually rare in more than a decade. It's also caused some bug reports as people were confused and/or didn't find a way to turn it off (there isn't one). Adding more NPCs each expansion is another task to take care of (maintenance cost), which may not be worth it in this case.

Let's try something new and completely different: Just remove it and see if anyone complains. It's easy to bring this functionality back if needed, in which case a setting to disable the tooltip addition should however be added alongside it.