Refactor: Replace holiday textures with calendar event IDs
Holiday events used to be detected via iconPath (texture names), which have later been replaced by artTextureID. There are now DBC tables that can be used to more easily detect holiday events based on their ID. Some events actually don't have a texture, especially if their status is "ONGOING". This is the case for the Greedy Emissary event, which means Rarity can't detect them while still using the old system.

The biggest drawback is that calendar events are instances with their own unique HolidayID, all of which share a HolidayNameID and HolidayDescriptionID. Using those would be preferable, but I haven't found a way to look them up ingame. This means that in the future, new instances (with different IDs) could be added. At that point, the addon database must be updated before they would be detected.

It's not great and I'll try to look into automating this when the time comes. For now, I just want to make it work because I don't have enough information to build a general solution, nor the time to learn all about calendar APIs and how they've changed over the years. I've confirmed this works on Classic as well as Retail, so it should be "good enough" for the time being.