Options: Merge the progress bar anchor and locked settings
Displaying the anchor as a visual indicator of the bar's locked status works better if there's no way to toggle each individually. If someone cares only about seeing the anchor, leaving the bar unlocked shouldn't be much of an issue.

This also frees up some UI space for other settings related to the progress bar, although that's a secondary concern.