UI: Refactor the processItem GameTooltip hook
Cuts down on repeated and in parts faulty logic that caused items to not always be processed as expected. Specifically, the Headless Horseman container only displayed the first item out of three.

The new approach should be flexible enough to also handle the unit tooltips, and hopefully work on Classic/without LibQTip with minimal changes as it's just a line buffer with a separate render call at the end.  I haven't touched those yet, however, so that's mostly a guess. Once the UI boilerplate is isolated it should also be testable. Unfortunately that isn't currently possible due to the dependency on LibBabble, so I'll leave it for later.